optical-touch-system/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
eventlet==0.33.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0